# api/openapi.yaml
openapi: 3.0.3
info:
  title: "Auth Service API"
  version: "1.0.0"
  description: |
    Authentication and user/business management service.
    
    Provides endpoints for user registration, login, token refresh,
    profile management, user administration and business resources.
    Protected endpoints require a Bearer JWT access token.
servers:
  - url: "http://localhost:8080"
  - url: "https://api.example.com"
paths:
...
  /api/v1/business:
    post:
      summary: Create a business
      description: Create a new business entry. Protected endpoint.
      tags: [Business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'

    get:
      summary: List businesses
      description: Return list of businesses (paginated optional).
      tags: [Business]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Business'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
  /api/v1/business/{id}:
    get:
      summary: Get business by ID
      description: Retrieve a business resource by ID.
      tags: [Business]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update business by ID
      description: Update business fields.
      tags: [Business]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessUpdate'
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete business by ID
      description: Permanently delete a business.
      tags: [Business]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Business deleted
        "404":
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Business:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: "b550e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: "Acme Corp"
        logo:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    BusinessCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        logo:
          type: string
          format: uri
        description:
          type: string
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string
          format: uri
        description:
          type: string
...
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
...
  responses:
    BadRequest:
      description: Validation failed or bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...